Remove build script and build via Make
Now if `make` is called after the project was already build, it will not be
rebuilt (unless `make -B` is used, or the ./build script is called
directly).

This also enables using `make -j` to fetch the repos and build them in
parallel, which can speed up the building a bit. Also add some comments
to Build.mak.