Change profiling unit from us to cycles and intoduce padding
This patch fixes the profiler-outptut form looking like this
```
5078127277490292274277443610704       54635     int ocean.sys.Epoll.Epoll.wait(ocean.sys.Epoll.epoll_event_t[], int)
```
to looking like this

```
     16384            680160372            680160372                41513     void test_prof.f2()
```

Because it is hard to get the cpu frequency under linux,
the profiler uses a hard-coded frequency (3mhz) when running linux.

Thereby introducing misleading information into the perforamnce-profile.

We are circumventing the problem by printing cycles instead of us,
and fix the ouput to include the necessary left-padding for the larger figures.